requirements.txt
google-cloud-bigquery
pandas
pyarrow
python-dotenv
